+++ packages/react/src/layout/Dynamic.js
@@ -1,2 +1,5 @@
 import React from 'react'
-export default ({ component: C = null, ...props }) => C && <C {...props} />
+let Dynamic = ({ component: C = null, ...props }) => C && <C {...props} />
+Dynamic.displayName = 'Dynamic'
+
+export default Dynamic