+++ packages/react/src/layout/LensInput.js
@@ -2,6 +2,9 @@ import React from 'react'
 import { observer } from 'mobx-react'
 import { value } from '../utils/actout'
 
-export default observer(({ lens, Input, ...x }) => (
+let LensInput = observer(({ lens, Input, ...x }) => (
   <Input {...value(lens)} {...x} />
 ))
+LensInput.displayName = 'LensInput'
+
+export default LensInput